requirements.txt
torch==2.4.0
torchvision==0.19.0
numpy
matplotlib
Pillow
tk
jsonschema
scikit-image
streamlit
opencv-python